// Frontend/WManager/src/app/model/newUserDto.ts
/**
 * OpenAPi WManager
 * OpenApi documentation for Spring Security
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface NewUserDto { 
    email?: string;
    firstName?: string;
    lastName?: string;
    password?: string;
}